+++ bot/src/test/java/edu/java/bot/ChatProcessorTest.java
@@ -10,6 +10,7 @@ import edu.java.bot.commands.TrackCommand;
 import edu.java.bot.commands.UnTrackCommand;
 import edu.java.bot.processor.ChatProcessor;
 import edu.java.bot.processor.DefaultChatProcessor;
+import edu.java.bot.user.InMemoryUserRepository;
 import edu.java.bot.user.InMemoryUserService;
 import edu.java.bot.user.UserService;
 import org.junit.jupiter.api.Test;
@@ -20,7 +21,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class ChatProcessorTest {
     @Test
     public void unknownCommandShouldReturnExpectedMessage() {
-        UserService userService = new InMemoryUserService();
+        UserService userService = new InMemoryUserService(new InMemoryUserRepository());
         ChatProcessor chatProcessor = new DefaultChatProcessor(List.of(
             new ListCommand(userService),
             new StartCommand(userService),
