+++ bot/src/test/java/edu/java/bot/InMemoryUserServiceTest.java
@@ -1,7 +1,9 @@
 package edu.java.bot;
 
+import edu.java.bot.user.InMemoryUserRepository;
 import edu.java.bot.user.InMemoryUserService;
 import edu.java.bot.exceptions.UserIsNotRegisteredException;
+import edu.java.bot.user.UserRepository;
 import edu.java.bot.user.UserService;
 import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -9,7 +11,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 public class InMemoryUserServiceTest {
     @Test
     public void listLinksOfUnregisteredUserShouldThrowException() {
-        UserService userService = new InMemoryUserService();
+        UserService userService = new InMemoryUserService(new InMemoryUserRepository());
 
         assertThatThrownBy(() -> userService.getTrackedLinks(0L))
             .isInstanceOf(UserIsNotRegisteredException.class);
